Revert "Revamp  `Manage Fleet` UI"
