+++ app/src/admin/java/dev/forcecodes/truckme/ui/fleet/BaseFleetAdapter.kt
@@ -15,6 +15,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import dev.forcecodes.truckme.R
 import dev.forcecodes.truckme.core.data.fleets.FleetUiModel
+import dev.forcecodes.truckme.core.domain.fleets.FleetStateUpdateMetadata
 import dev.forcecodes.truckme.core.util.then
 import dev.forcecodes.truckme.databinding.FleetItemBinding
 import dev.forcecodes.truckme.extensions.bindImageWith
@@ -56,7 +57,9 @@ abstract class BaseFleetAdapter<T : FleetUiModel>(
     val binding = FleetItemBinding.inflate(
       LayoutInflater.from(parent.context),
       parent, false
-    )
+    ).apply {
+      fleetType.text = parent.context.getString(adapterTitle)
+    }
     return FleetViewHolder(binding)
   }
 
@@ -75,7 +78,9 @@ abstract class BaseFleetAdapter<T : FleetUiModel>(
     with(holder.binding) {
       fleetContainer.setOnClickListener { onViewHolderCreated(item) }
       moreButton.setOnClickListener { popUpDelete(it, position, item) }
+      fleetType.isVisible = position == 0 && itemCount > 0
     }
+
   }
 
   private fun popUpDelete(view: View, position: Int, item: T) {
@@ -88,8 +93,8 @@ abstract class BaseFleetAdapter<T : FleetUiModel>(
           onDeleteFleet(getItem(position).id)
         }
         if (it.itemId == R.id.delivery_state) {
-          val invertedState = !item.isActive
-          onChangeActiveState(item.id, invertedState)
+            val invertedState = !item.isActive
+            onChangeActiveState(item.id, invertedState)
         }
         return@setOnMenuItemClickListener true
       }
