--- app/src/admin/java/dev/forcecodes/truckme/ui/fleet/FleetPageType.kt
@@ -1,6 +1,0 @@
-package dev.forcecodes.truckme.ui.fleet
-
-enum class FleetPageType {
-  DRIVER,
-  VEHICLE
-}
