+++ app/src/admin/res/layout/fleet_item.xml
@@ -7,6 +7,17 @@
     android:orientation="vertical"
     >
 
+    <TextView
+        android:id="@+id/fleet_type"
+        style="@style/TextAppearance.AppCompat.Headline"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/spacing_extra_small"
+        android:paddingTop="@dimen/spacing_medium"
+        android:paddingBottom="@dimen/spacing_medium"
+        android:textStyle="bold"
+        android:visibility="gone" />
+
     <androidx.cardview.widget.CardView
         style="@style/ClickableButtonStyle"
         android:id="@+id/fleet_container"
