--- app/src/admin/res/layout/fragment_fleet_pages.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    >
-
-  <androidx.recyclerview.widget.RecyclerView
-      android:paddingTop="@dimen/spacing_small"
-      android:id="@+id/fleet_list"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:layout_marginEnd="@dimen/spacing_small"
-      android:layout_marginStart="@dimen/spacing_small"
-      android:clipToPadding="false"
-      android:nestedScrollingEnabled="false"
-      android:paddingBottom="@dimen/spacing_normal"
-      app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-      tools:listitem="@layout/fleet_item"
-      />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
